Added catalog option to OPAT. Removed librarian for now.
