+++ src/opat.ts
@@ -11,6 +11,8 @@
  * Example commands:
  *   npx ts-node src/opat.ts validate -f tests/examples/valid.yaml # Output: Valid!
  *   npx ts-node src/opat.ts validate -f tests/examples/invalid-basic.yaml # Output: Invalid: ...
+ *   npx ts-node src/opat.ts validate -f tests/examples/valid.yaml -cf catalog/wcag2-catalog.yaml # Output: Valid!
+ *   npx ts-node src/opat.ts validate -f tests/examples/invalid-criteria.yaml -cf catalog/wcag2-catalog.yaml # Output: Invalid: ...
  */
 
 import { validateOPAT } from "./validateOPAT";
@@ -18,6 +20,7 @@ import yargs from "yargs";
 import fs from "fs";
 import yaml from "js-yaml";
 import { ValidatorResult } from "./ValidatorResult";
+import { validateOPATCatalogValues } from "./validateOPATCatalogValues";
 
 const argv = yargs
   .command("validate", "Validate OPAT content", function (yargs) {
@@ -27,6 +30,12 @@ const argv = yargs
         description: "Content filename",
         demandOption: true,
         alias: "f",
+      },
+      catalogFile: {
+        type: "string",
+        description: "Catalog filename",
+        demandOption: false,
+        alias: "cf",
       },
     });
   })
@@ -40,6 +49,18 @@ if (fs.existsSync(argv.file)) {
   try {
     const data = yaml.load(fs.readFileSync(argv.file).toString());
     result = validateOPAT(data, schema);
+    // Validate OPAT against provided catalog.
+    if (result.result && argv.catalogFile && fs.existsSync(argv.catalogFile)) {
+      try {
+        const catalog = yaml.load(fs.readFileSync(argv.catalogFile).toString());
+        result = validateOPATCatalogValues(data, catalog);
+      } catch {
+        result = {
+          result: false,
+          message: "Invalid: catalog file is not in YAML format",
+        };
+      }
+    }
   } catch {
     result = {
       result: false,
