+++ tests/opat-cli.test.ts
@@ -20,7 +20,7 @@ describe("Validate OPAT CLI", () => {
     });
   });
 
-  it("when passed not a YAML file should return invalid message", () => {
+  it("when passed an invalid YAML file should return invalid message", () => {
     const invalid = spawn(cmd, options.concat("README.md"));
     const chunks = [];
 
@@ -32,6 +32,24 @@ describe("Validate OPAT CLI", () => {
       const output = Buffer.concat(chunks).toString();
 
       expect(output).to.equal("Invalid: file is not in YAML format\n");
+    });
+  });
+
+  it("when passed an invalid YAML catalog file should return invalid message", () => {
+    const invalid = spawn(
+      cmd,
+      options.concat("tests/examples/valid.yaml", "--cf", "README.md")
+    );
+    const chunks = [];
+
+    invalid.stderr.on("data", (chunk) => {
+      chunks.push(chunk);
+    });
+
+    invalid.stderr.on("end", () => {
+      const output = Buffer.concat(chunks).toString();
+
+      expect(output).to.equal("Invalid: catalog file is not in YAML format\n");
     });
   });
 
@@ -57,8 +75,47 @@ describe("Validate OPAT CLI", () => {
     });
   });
 
+  // it("when passed an invalid criteria example should return invalid message", () => {
+  //   const invalid = spawn(cmd, options.concat("tests/examples/invalid-criteria.yaml"));
+  //   const chunks = [];
+  //
+  //   invalid.stderr.on("data", (chunk) => {
+  //     chunks.push(chunk);
+  //   });
+  //
+  //   invalid.stderr.on("end", () => {
+  //     const output = Buffer.concat(chunks).toString();
+  //
+  //     expect(output).to.equal(
+  //       "Invalid: criteria num '1.2.2' is not included in chapter 'Success Criteria, Level A'"
+  //     );
+  //   });
+  // });
+
   it("when passed a valid file should return valid message", () => {
     const valid = spawn(cmd, options.concat("tests/examples/valid.yaml"));
+    const chunks = [];
+
+    valid.stdout.on("data", (chunk) => {
+      chunks.push(chunk);
+    });
+
+    valid.stdout.on("end", () => {
+      const output = Buffer.concat(chunks).toString();
+
+      expect(output).to.equal("Valid!\n");
+    });
+  });
+
+  it("when passed a valid file and valid catalog file should return valid message", () => {
+    const valid = spawn(
+      cmd,
+      options.concat(
+        "tests/examples/valid.yaml",
+        "--cf",
+        "catalog/wcag2-catalog.yaml"
+      )
+    );
     const chunks = [];
 
     valid.stdout.on("data", (chunk) => {
